requirements.txt
dynaconf==2.2.1
Flask-SQLAlchemy>=2.4.1
Flask==1.1.1
SQLAlchemy-serializer>=1.3.4.2
flask_script==2.0.6
flask_migrate==2.5.3
psycopg2>=2.8.5
marshmallow==3.6.1
PyJWT==1.7.1
flask-restplus>=0.12.0
gunicorn==20.0.4
Flask-SocketIO==5.0.1
flask-restplus>=0.12.0
Werkzeug==0.16.1
flask-simplelogin==0.0.7
WTForms==2.3.3
qrcode==6.1
Flask-Session==0.3.2
python-dotenv==0.17.0
Flask-QRcode==3.0.0
Flask-Login==0.5.0
Flask-Cors==3.0.10